+++ rbx_types/src/variant.rs
@@ -160,9 +160,9 @@ mod serde_test {
     fn non_human() {
         let vec2 = Variant::Vector2(Vector2::new(5.0, 7.0));
 
-        let ser = bincode::serialize(&vec2).unwrap();
+        let ser = rmp_serde::to_vec(&vec2).unwrap();
 
-        let de: Variant = bincode::deserialize(&ser).unwrap();
+        let de: Variant = rmp_serde::from_slice(&ser).unwrap();
         assert_eq!(de, vec2);
     }
 }